requirements.txt
scipy
numpy
matplotlib
opencv_python==4.1.2.30
torch==1.4.0
torchvision==0.5.0
tqdm==4.60.0
Pillow==8.2.0
h5py==2.10.0
face_recognition
